arm: tegra: power: clear pmc wake status register on suspend
PMC register hold the reason for wakeup from LP0. The keypad driver resume the
device from the perspective of the User if the Power Key event is sent in the
register. Reset the register before going to suspend so that the status won't
get carried to the next wake time as in this case of LP1.

Bug 909191
Bug 913110

Change-Id: I0f23682ff4b5017884896e149d42461d3041c5d6
Signed-off-by: Sanjay Singh Rawat <srawat@nvidia.com>
Reviewed-on: http://git-master/r/70134
Reviewed-by: Jihoon Bang <jbang@nvidia.com>
Tested-by: Jihoon Bang <jbang@nvidia.com>
Reviewed-by: Ryan Wong <ryanw@nvidia.com>