DTV: dma: fixed dma burst size issue DTV xmit
Since dtv interface was designed based on SPI bus, it shares the same
dma configuration with SPI bus. However, it is not proper because DTV
interface has to transmit data in 4 words long constantly. The patch
added an option in tegra_dma_req to set burst size to be fixed 4 words.

fixed Bug 910227

Change-Id: I1436f0c8d108dd39edc57ae4c9cb750d9574b62c
Signed-off-by: Adam Jiang <chaoj@nvidia.com>
Reviewed-on: http://git-master/r/75509
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: David Schalig <dschalig@nvidia.com>
Reviewed-by: Laxman Dewangan <ldewangan@nvidia.com>