ARM: tegra: configs: Enable MTD partition and MTD CHAR
Allow generic configuration of the MTD partition tables via the kernel
command line.
Also provide a character device for each MTD device present in
the system, allowing the user to read and write directly to the
memory chips, and also use ioctl() to obtain information about
the device, or to erase parts of it.

Bug: 923135
Change-Id: If6ae96da379e9fdd89c28e476054ec3352373d47
Signed-off-by: Preetham Chandru R <pchandru@nvidia.com>
Reviewed-on: http://git-master/r/75921
Reviewed-by: Rohan Somvanshi <rsomvanshi@nvidia.com>
Tested-by: Rohan Somvanshi <rsomvanshi@nvidia.com>