sched: domain: Convert freeing old root domain to use RCU
Bug 923063

Instead of calling synchronize_sched(), directly invoke
call_rcu_sched() which will not block the cpu hotplug operation but
will call free_rootdomain when the rcu process is complete. This is
safe because the root domains are allocated dynamically so any new
readers will have pointers to the new data and the new data does not
overwrite the old data. This is a partial backport from upstream.

  dce840a08702bd13a9a186e07e63d1ef82256b5e
  Author: Peter Zijlstra <a.p.zijlstra@chello.nl>
  Date:   Thu Apr 7 14:09:50 2011 +0200

  sched: Dynamically allocate sched_domain/sched_group data-structures

  Instead of relying on static allocations for the sched_domain and
  sched_group trees, dynamically allocate and RCU free them.

  Signed-off-by: Peter Zijlstra <a.p.zijlstra@chello.nl>

Change-Id: I133e15fba6360416e1bd85b1bd5d7f219ef403a1
Signed-off-by: Diwakar Tundlam <dtundlam@nvidia.com>
Reviewed-on: http://git-master/r/84421
Reviewed-by: Varun Colbert <vcolbert@nvidia.com>
Tested-by: Varun Colbert <vcolbert@nvidia.com>