video: tegra: dc: acquire window locks in an order to avoid a deadlock
All window locks are grabbed while performing dc window updates. Currently, no
particular locking order is followed for these locks. If user provided windows
are not in order, this leads to deadlock due to race between flip ioctl and
dc underflow reset worker.

Now on all window locks are acquired in an order as below
1. window A i.e. index 0
2. window B i.e. index 1
3. window C i.e. index 2

And unlocked in the reverse order
1. window C i.e. index 2
2. window B i.e. index 1
3. window A i.e. index 0

Bug 936545

Change-Id: I487f1403f63e0276b9ea2f9e782c041e260f164f
Signed-off-by: Nitin Kumbhar <nkumbhar@nvidia.com>
Reviewed-on: http://git-master/r/83107
Reviewed-by: Bharat Nihalani <bnihalani@nvidia.com>
Tested-by: Bharat Nihalani <bnihalani@nvidia.com>
Reviewed-by: Robert Morell <rmorell@nvidia.com>