arm: tegra: cardhu: increase mc outstanding reqs for lower ram freqs
The number of outstanding memory transactions is limited by a setting
in MC. This leads to dc underflows which cause flickers on lcd panel.
Increase the limits to optimal values which don't show dc underflows.

Currently, the cap on outstanding requests for a frequency is calculated
by linearly scaling up values for frequencies keeping minimum value at
0x08. An exception has to be made to resolve dc underflows and
lcd flickers.

For cardhu, the lower ram frequencies are 25.5MHz, 51MHz and 102MHz.
So increase minimum value to 0x10 and set 0x18 for 102MHz as an optimal
value with which there are no dc underflows.

Memory tables of Hynix-1GB, Hynix-2GB and Samsung-2GB memory types are
updated with this change.

Bug 932113
Bug 946316

Change-Id: I823cd47958a70843662ff836ed7beb150f16f692
Signed-off-by: Nitin Kumbhar <nkumbhar@nvidia.com>
Reviewed-on: http://git-master/r/87230
Reviewed-by: Bharat Nihalani <bnihalani@nvidia.com>
Reviewed-by: Edward Ahn <eahn@nvidia.com>
Reviewed-by: Jon Mayo <jmayo@nvidia.com>